refactor(enhanceConfigure): export it from InstantSearch.js
